A5/cloud/scalability/requirements.txt
pymongo
datetime
mysql-connector